Move code to preferred layout
