+++ orientdb/src/main/scala/akka/stream/alpakka/orientdb/scaladsl/OrientDBSource.scala
@@ -6,6 +6,7 @@ package akka.stream.alpakka.orientdb.scaladsl
 
 import akka.NotUsed
 import akka.stream.alpakka.orientdb._
+import akka.stream.alpakka.orientdb.impl.{MessageReader, OrientDBSourceStage}
 import akka.stream.scaladsl.Source
 import com.orientechnologies.orient.core.record.impl.ODocument
