+++ orientdb/src/test/resources/application.conf
@@ -1,0 +1,5 @@
+akka {
+  loggers = ["akka.testkit.TestEventListener", "akka.event.slf4j.Slf4jLogger"]
+  logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
+  loglevel = "DEBUG"
+}
