+++ orientdb/src/test/resources/logback-test.xml
@@ -1,0 +1,17 @@
+<configuration>
+    <appender name="FILE" class="ch.qos.logback.core.FileAppender">
+        <file>target/orientdb.log</file>
+        <append>false</append>
+        <encoder>
+            <pattern>%d{ISO8601} %-5level [%thread] [%logger{36}]  %msg%n</pattern>
+        </encoder>
+    </appender>
+
+
+    <logger name="akka" level="info"/>
+    <logger name="akka.stream.alpakka" level="debug"/>
+
+    <root level="debug">
+        <appender-ref ref="FILE" />
+    </root>
+</configuration>
